Add comments and organize the C++ header into the main header plus a separate one of the inline methods.
